+++ .nuke/build.schema.json
@@ -118,10 +118,6 @@
   "allOf": [
     {
       "properties": {
-        "BenchmarkFilter": {
-          "type": "string",
-          "description": "Filter for BenchmarkDotNet - Default is '*'"
-        },
         "Configuration": {
           "type": "string",
           "description": "Configuration to build - Default is 'Debug' (local) or 'Release' (server)",
